% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BARMAJournalHydrology2024-package.R
\docType{package}
\name{BARMAJournalHydrology2024-package}
\alias{BARMAJournalHydrology2024}
\alias{BARMAJournalHydrology2024-package}
\title{BARMAJournalHydrology2024: Test Inferences and Link Function Selection in Dynamic Beta Modeling of Seasonal Hydro-environmental Time Series with Temporary Abnormal Regimes}
\description{
Provides R code and data to replicate the analysis in Costa, Cribari-Neto, and Scher (2024) \doi{10.1016/j.jhydrol.2024.131489}. The paper considers the modeling and forecasting of hydro-environmental time series subject to seasonal fluctuations and prolonged droughts using a class of dynamic beta (BARMA) models. The code addresses the accuracy of hypothesis tests and introduces a new model selection criterion ('the scout score') for selecting the link function of the model.
}
\seealso{
Useful links:
\itemize{
  \item \doi{10.1016/j.jhydrol.2024.131489}
  \item Report bugs at \url{https://github.com/everton-da-costa/BARMAJournalHydrology2024/issues}
}

}
\author{
\strong{Maintainer}: Everton Costa \email{everto.cost@gmail.com} (\href{https://orcid.org/0000-0001-7580-2639}{ORCID}) (Conceptualization, Formal analysis, Software, Validation, Visualization) [copyright holder]

Authors:
\itemize{
  \item Francisco Cribari-Neto (\href{https://orcid.org/0000-0002-5909-6698}{ORCID}) (Conceptualization, Formal analysis, Funding acquisition, Investigation, Methodology, Project administration, Resources, Supervision, Validation, Writing)
  \item Vinicius Scher (\href{https://orcid.org/0000-0003-0406-0265}{ORCID}) (Formal analysis, Investigation, Methodology, Resources, Software, Visualization)
}

}
\keyword{internal}
