% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_classical_tests.R
\name{compute_classical_tests}
\alias{compute_classical_tests}
\title{Compute Classical Hypothesis Tests}
\usage{
compute_classical_tests(
  unrestricted_model,
  loglik_restr_naive,
  loglik_restr_m0,
  loglik_restr_mfun,
  score_vec_naive,
  score_vec_mfun,
  matrix_inf_naive,
  matrix_inf_mfun,
  rest,
  num_rest
)
}
\arguments{
\item{unrestricted_model}{The fitted unrestricted model object.}

\item{loglik_restr_naive}{Log-likelihood of the restricted model (using a_N).}

\item{loglik_restr_m0}{Adjusted log-likelihood of the restricted model.}

\item{loglik_restr_mfun}{Log-likelihood of the restricted model (using a_NN).}

\item{score_vec_naive}{Score vector from the restricted model (using a_N).}

\item{score_vec_mfun}{Score vector from the restricted model (using a_NN).}

\item{matrix_inf_naive}{Information matrix from the restricted model
(using a_N).}

\item{matrix_inf_mfun}{Information matrix from the restricted model
(using a_NN).}

\item{rest}{A vector of indices for the restricted parameters.}

\item{num_rest}{The number of restricted parameters (degrees of freedom).}
}
\value{
A list (`z_test`) containing the results of all computed tests.
}
\description{
Internal function to compute Wald, LR, and Rao Score tests.
  Methodological details based on Costa et al. (2024) J.Hydrol. 638, 131489.
}
\keyword{internal}
