% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inf_matrix_arma.R
\name{inf_matrix_arma}
\alias{inf_matrix_arma}
\title{Compute Fisher Information Matrix for BARMA Model (Auxiliary)}
\usage{
inf_matrix_arma(y, ar, ma, link, alpha = 0, varphi = 0, theta = 0, phi = 0)
}
\arguments{
\item{link}{A string specifying the link function (e.g., "logit", "probit",
"cloglog", "loglog"). This defines the transformation
\eqn{\eta = g(\mu)} linking the mean response \eqn{\mu \in (0,1)}
to the linear predictor \eqn{\eta}. It's used to derive the link
function \eqn{g(.)}, the inverse link (mean function)
\eqn{\mu = g^{-1}(\eta)}, and its derivative \eqn{d\mu/d\eta}.}

\item{alpha}{Scalar, the intercept parameter.}

\item{varphi}{Numeric vector, AR coefficients (\eqn{\varphi}).}

\item{theta}{Numeric vector, MA coefficients (\eqn{\theta}).}

\item{phi}{Scalar, the precision parameter (\eqn{\phi}).}
}
\value{
A matrix representing the Fisher Information Matrix for the
        parameters (\eqn{\alpha}, \eqn{\varphi}, \eqn{\theta}, \eqn{\phi}).
}
\description{
Calculates the expected Fisher Information Matrix for a Beta Autoregressive
Moving Average (BARMA) model. This is an internal helper function.
The matrix is derived based on the methodology in Rocha & Cribari-Neto
(2009, 2017) and is essential for statistical inference, e.g., for
calculating standard errors or for constructing Wald and Score tests
as discussed in Costa et al. (2024).
}
\details{
This function uses the time series `y` (values in (0,1)) and a `link`
string (e.g., "logit") to internally derive `linkfun` (g(.)), `linkinv`
(\eqn{\mu = g^{-1}(\eta)}), and `mu.eta` (\eqn{d\mu/d\eta}). The transformed
series `ynew = linkfun(y)` is then used.
It computes components of the Fisher Information Matrix for parameters
\eqn{\alpha, \varphi, \theta, \phi}. `errorhat` are on the
predictor scale.
}
\references{
Rocha, A.V., Cribari-Neto, F. (2009). Beta autoregressive moving
average models. TEST, 18(3), 529--545.
Costa, E., Cribari-Neto, F., Scher, V.T. (2024). Test inferences and
link function selection in dynamic beta modeling... Journal of Hydrology,
638, 131489.
}
\keyword{internal}
