% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{itaparica_df}
\alias{itaparica_df}
\title{Monthly Useful Volume of the Itaparica Reservoir}
\format{
A data frame with 2 variables:
\describe{
  \item{time}{Month and year of observation (yearmon format)}
  \item{y}{Useful volume as a proportion (between 0 and 1)}
}
}
\source{
Brazilian National Electric System Operator (ONS)
}
\usage{
itaparica_df
}
\description{
A time series dataset containing the monthly useful volume, expressed as a
proportion (ranging from 0 to 1), of the Itaparica hydroelectric power
plant reservoir. The reservoir is part of the São Francisco River Basin (SFRB)
in Brazil. The data covers a period marked by seasonal fluctuations and
prolonged droughts.
}
\keyword{datasets}
