% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{itaparica_ts}
\alias{itaparica_ts}
\title{Itaparica Reservoir Monthly Useful Volume Time Series}
\format{
A time series object (`ts`) with 301 monthly observations from
January 1999 to January 2024.
}
\source{
Brazilian National Electric System Operator (ONS)
}
\usage{
itaparica_ts
}
\description{
Monthly useful volume data from the Itaparica reservoir as a time series object,
starting from January 1999.
}
\keyword{datasets}
