% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_terms_ar.R
\name{loglik_terms_ar}
\alias{loglik_terms_ar}
\title{Compute Log-Likelihood Terms for Beta AR Model (Auxiliary)}
\usage{
loglik_terms_ar(y, ar, link, alpha = 0, varphi = 0, phi = 0)
}
\arguments{
\item{ar}{Numeric vector, AR orders to be included in the model.}

\item{link}{A string specifying the link function (e.g., "logit", "probit",
"cloglog", "loglog"). This defines the transformation
\eqn{\eta = g(\mu)} linking the mean response \eqn{\mu \in (0,1)}
to the linear predictor \eqn{\eta}. It's used to derive the link
function \eqn{g(.)}, the inverse link (mean function)
\eqn{\mu = g^{-1}(\eta)}, and its derivative \eqn{d\mu/d\eta}.}

\item{alpha}{Scalar, the intercept parameter.}

\item{varphi}{Numeric vector, AR coefficients (\eqn{\varphi}) corresponding
to lags in `ar`.}

\item{phi}{Scalar, the precision parameter (\eqn{\phi}).}
}
\value{
A numeric vector of individual log-likelihood terms.
}
\description{
Calculates the individual log-likelihood terms for a Beta Autoregressive
(BAR) model. This is an internal helper function.
These terms can be summed for the total log-likelihood.
}
\details{
This function takes the time series `y` (with values in the interval (0,1))
and a `link` specification (e.g., "logit", "probit", "cloglog") as input.
It internally generates `linkfun` (g(.)) and `linkinv` (the inverse of g(.),
also known as the mean function \eqn{\mu = g^{-1}(\eta)}) using the
provided `link` argument. The transformed time series `ynew = linkfun(y)`
and the `linkinv` function are then used to calculate the log-likelihood
terms.`error` are on the predictor scale.
This function is particularly useful for tests like LR2 in Costa et al.
(2024), which may require summing specific subsets of log-likelihood
terms from a null model.
}
\keyword{internal}
