% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_terms_ma.R
\name{loglik_terms_ma}
\alias{loglik_terms_ma}
\title{Compute Log-Likelihood Terms for Beta MA Model (Auxiliary)}
\usage{
loglik_terms_ma(y, ma, link, alpha = 0, theta = 0, phi = 0)
}
\arguments{
\item{ma}{Numeric vector, MA orders to be included in the model.}

\item{alpha}{Scalar, the intercept parameter.}

\item{theta}{Numeric vector, MA coefficients (\eqn{\theta}) corresponding
to lags in `ma`.}

\item{phi}{Scalar, the precision parameter (\eqn{\phi}).}
}
\value{
A numeric vector of individual log-likelihood terms.
}
\description{
Calculates the individual log-likelihood terms for a Beta Moving Average
(BMA) model. This is an internal helper function.
}
\details{
#' This function takes the time series `y` (with values in the interval
(0,1)) and a `link` specification (e.g., "logit", "probit", "cloglog") as
input. It internally generates `linkfun` (g(.)) and `linkinv` (the inverse
of g(.), also known as the mean function \eqn{\mu = g^{-1}(\eta)}) using the
provided `link` argument. The transformed time series `ynew = linkfun(y)`
and the `linkinv` function are then used to calculate the log-likelihood
terms. `error` are on the predictor scale. This function is particularly
useful for tests like LR2 in Costa et al. (2024), which may require summing
 specific subsets of log-likelihood terms from a null model.
}
\keyword{internal}
