% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_vector_arma.R
\name{score_vector_arma}
\alias{score_vector_arma}
\title{Compute Score Vector for BARMA Model (Auxiliary)}
\usage{
score_vector_arma(y, ar, ma, link, alpha = 0, varphi = 0, theta = 0, phi = 0)
}
\arguments{
\item{link}{A string specifying the link function (e.g., "logit", "probit",
"cloglog", "loglog"). This defines the transformation
\eqn{\eta = g(\mu)} linking the mean response \eqn{\mu \in (0,1)}
to the linear predictor \eqn{\eta}. It's used to derive the link
function \eqn{g(.)}, the inverse link (mean function)
\eqn{\mu = g^{-1}(\eta)}, and its derivative \eqn{d\mu/d\eta}.}

\item{alpha}{Scalar, the intercept parameter.}

\item{varphi}{Numeric vector, AR coefficients (\eqn{\varphi}).}

\item{theta}{Numeric vector, MA coefficients (\eqn{\theta}).}

\item{phi}{Scalar, the precision parameter (\eqn{\phi}).}
}
\value{
A numeric vector representing the score for parameters
        (\eqn{\alpha}, \eqn{\varphi}, \eqn{\theta}, \eqn{\phi}).
}
\description{
Calculates the score vector (gradient of the log-likelihood) for a BARMA
model. This is an internal helper function.
The score vector is used in ML estimation and for Score tests.
}
\details{
This function uses the time series `y` (values in (0,1)) and a `link`
string (e.g., "logit") to internally derive `linkfun` (g(.)), `linkinv`
(\eqn{\mu = g^{-1}(\eta)}), and `mu.eta` (\eqn{d\mu/d\eta}). The transformed
series `ynew = linkfun(y)` is then used.
It computes components of the Score Vector for parameters
\eqn{\alpha, \varphi, \theta, \phi}. `error` are on the
predictor scale. `ystar` is `logit(y1)` and `mustar` involves digamma
functions.
}
\keyword{internal}
