% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{tres_marias}
\alias{tres_marias}
\title{Monthly Useful Volume of the Três Marias Reservoir}
\format{
A time series object (`ts`) with 301 monthly observations from
January 1999 to January 2024.
}
\source{
The data was obtained from the Brazilian National Electric System Operator
(Operador Nacional do Sistema Elétrico - ONS), available at
\url{https://www.ons.org.br}.
}
\description{
A time series dataset containing the monthly useful volume, expressed as a
proportion (ranging from 0 to 1), of the Três Marias hydroelectric power
plant reservoir. The reservoir is part of the São Francisco River Basin (SFRB)
in Brazil. The data covers a period marked by seasonal fluctuations and
prolonged droughts.
}
